MICRO: Generate Parameters for Landscape AquaticSite

This function generates a named list of parameters to initialize all AquaticSite objects on a MICRO Landscape.

Landscape.Aqua.Parameters(nAqua, siteXY, module, modulePars, searchWt = NULL,
  haz = 0)

Arguments

nAqua
number of aquatic habitats

siteXY
two element list of x and y coordinates of aquatic habitats

module
character

"emerge": initialize parameters for Emerge module of Aquatic Ecology

"EL4P": initialize parameters for EL4P module of Aquatic Ecology

modulePars
additional list of named parameters to be passed to Aquatic Ecology module specific parameter generating functions

Emerge: see for deatils makeLambda_Micro

EL4P:

searchWt
vector of searchWt (if NULL initialize to Gamma(1,1) distribution)

haz
mean value of landing hazards (if != 0 use betaRootA to find alpha parameter of beta distribution to give that mean value and produce Beta distributed hazards)

Value

return a list