Initialize MosquitoPop Parameters for MacroMosquitoPop

make a list of pars for MacroMosquitoPop

MACRO.MosquitoPop.Parameters(N, EIP = 10, M_density = 10, p = 0.9,
  f = 0.3, Q = 0.9, v = 20, maxEIP = 30, psi = NULL)

Arguments

N
number of patches

EIP
entomological incubation period

M_density
either a vector of mosquito densities equal to N or a single value for all patches

p
daily survival probability

f
feeding rate (average number of daily blood meals per capita)

Q
human blood index (propensity to feed on humans)

v
daily egg laying rate

maxEIP
maximum length (in days) to track mosquito progression through EIP

psi
movement matrix for rough diffusion mosquito movement, if NULL mosquitoes do not move from patch to patch

Value

return a list MacroMosquitoPop_PAR

Examples

MACRO.MosquitoPop.Parameters(N = 5)

#> $p
#> [1] 0.9
#> 
#> $f
#> [1] 0.3
#> 
#> $Q
#> [1] 0.9 0.9 0.9 0.9 0.9
#> 
#> $v
#> [1] 20
#> 
#> $EIP
#> [1] 10
#> 
#> $maxEIP
#> [1] 30
#> 
#> $psi
#> NULL
#> 
#> $M_density
#> [1] 10 10 10 10 10
#>