MICRO: Generate Parameters for MicroTile Object

This function is a specific instantiation of a generic system to generate parameters for a chosen microsimulation Tile. Any user-specified function can be written to generate parameters, as long as the return list is in the same format.

MICRO.Tile.Parameters(nFeed, nAqua, pointGen = "poisson", module, modulePars,
  hazV = 0, hazW = 0, hazI = 0, haz = 0, searchFeed = NULL,
  searchAqua = NULL, enterP = NULL, xLim = c(0, 1), yLim = c(0, 1),
  aquaSD = 0.025, hhSize = 7, hhMin = 2, bWeight = NULL, ...)

Arguments

nFeed
number of feeding sites

nAqua
number of aquatic habitats

pointGen
character to select spatial point pattern generation function

"poisson": pointsPoisson

"clustered": pointsClustered

"overdispersed": pointsOverdispersed

"lattice": pointsLattice

module
character

"emerge": initialize parameters for Emerge module of Aquatic Ecology

"EL4P": initialize parameters for EL4P module of Aquatic Ecology

modulePars
additional list of named parameters to be passed to Aquatic Ecology module specific parameter generating functions

Emerge: for details see makeLambda_Micro

EL4P:

hazV
mean value for feeding site vegetation landing hazard (if 0 it is set to 0 for all sites)

hazW
mean value for feeding site outside wall landing hazard (if 0 it is set to 0 for all sites)

hazI
mean value for feeding site indoor wall landing hazard (if 0 it is set to 0 for all sites)

haz
mean value for aquatic habitat landing hazard (if 0 it is set to 0 for all sites)

searchFeed
vector of searchWt for feeding sites (if NULL initialize to Gamma(1,1) distribution)

searchAqua
vector of searchWt for aquatic habitats (if NULL initialize to Gamma(1,1) distribution)

enterP
vector of house entry probabilities or single numeric value for all sites (if NULL initialize to Beta(9,1) distribution)

xLim
x-axis bounds for simulated points

yLim
y-axis bounds for simulated points

aquaSD
standard deviation of aquatic habitat scatter around feeding sites

hhSize
average number of humans at feeding sites

hhMin
minimum number of humans at feeding sites

bWeight
numeric value of biting weights on Human (if NULL biting weights are Gamma(1,1) distributed)

...
additional named arguments for pointGen()

Value

a named list of parameters

Landscape_PAR: see Landscape.Parameters for details

HumanPop_PAR: see HumanPop.Parameters for details