MACRO: Generate Seasonal Emergence (Lambda) for Emerge model of Aquatic Ecology

Generate lambda for all patches.

makeLambda_Macro(aquaPars)

Arguments

aquaPars
a list of the following structure

lambda: vector of length equal to number of patches MacroPatch where each element is the number of emerging adult females per human per day averaged over one year (required)

lambdaWeight: vector of weights applied to each site (if not specified or set to NULL initialize to Gamma(1,1) distribution)

offset: vector of seasonal offsets in peak emergence applied to each site (if not specified or set to NULL initialize to 0 for all sites)

Value

list lambda where each element is the daily emergence for that MacroPatch

Examples

makeLambda_Macro(aquaPars = list(lambda=c(5,10,15)))

#> [[1]]
#>   [1] 2.541509e+00 2.584504e+00 2.627473e+00 2.670404e+00 2.713284e+00
#>   [6] 2.756101e+00 2.798841e+00 2.841492e+00 2.884042e+00 2.926477e+00
#>  [11] 2.968786e+00 3.010955e+00 3.052972e+00 3.094825e+00 3.136502e+00
#>  [16] 3.177989e+00 3.219275e+00 3.260347e+00 3.301194e+00 3.341803e+00
#>  [21] 3.382162e+00 3.422259e+00 3.462082e+00 3.501620e+00 3.540860e+00
#>  [26] 3.579792e+00 3.618403e+00 3.656682e+00 3.694618e+00 3.732200e+00
#>  [31] 3.769416e+00 3.806256e+00 3.842708e+00 3.878762e+00 3.914406e+00
#>  [36] 3.949632e+00 3.984427e+00 4.018782e+00 4.052686e+00 4.086130e+00
#>  [41] 4.119103e+00 4.151596e+00 4.183600e+00 4.215103e+00 4.246099e+00
#>  [46] 4.276576e+00 4.306527e+00 4.335942e+00 4.364812e+00 4.393129e+00
#>  [51] 4.420885e+00 4.448071e+00 4.474680e+00 4.500703e+00 4.526133e+00
#>  [56] 4.550961e+00 4.575182e+00 4.598787e+00 4.621770e+00 4.644124e+00
#>  [61] 4.665842e+00 4.686918e+00 4.707345e+00 4.727118e+00 4.746230e+00
#>  [66] 4.764676e+00 4.782451e+00 4.799549e+00 4.815965e+00 4.831695e+00
#>  [71] 4.846733e+00 4.861075e+00 4.874717e+00 4.887655e+00 4.899885e+00
#>  [76] 4.911404e+00 4.922207e+00 4.932293e+00 4.941657e+00 4.950297e+00
#>  [81] 4.958210e+00 4.965395e+00 4.971849e+00 4.977570e+00 4.982556e+00
#>  [86] 4.986806e+00 4.990319e+00 4.993093e+00 4.995129e+00 4.996424e+00
#>  [91] 4.996979e+00 4.996794e+00 4.995869e+00 4.994204e+00 4.991799e+00
#>  [96] 4.988655e+00 4.984773e+00 4.980155e+00 4.974801e+00 4.968713e+00
#> [101] 4.961894e+00 4.954345e+00 4.946067e+00 4.937065e+00 4.927340e+00
#> [106] 4.916895e+00 4.905734e+00 4.893859e+00 4.881274e+00 4.867984e+00
#> [111] 4.853991e+00 4.839301e+00 4.823916e+00 4.807843e+00 4.791085e+00
#> [116] 4.773648e+00 4.755537e+00 4.736757e+00 4.717314e+00 4.697213e+00
#> [121] 4.676461e+00 4.655063e+00 4.633026e+00 4.610357e+00 4.587062e+00
#> [126] 4.563148e+00 4.538623e+00 4.513492e+00 4.487765e+00 4.461448e+00
#> [131] 4.434550e+00 4.407078e+00 4.379040e+00 4.350445e+00 4.321302e+00
#> [136] 4.291618e+00 4.261403e+00 4.230665e+00 4.199415e+00 4.167660e+00
#> [141] 4.135410e+00 4.102676e+00 4.069466e+00 4.035791e+00 4.001660e+00
#> [146] 3.967083e+00 3.932072e+00 3.896636e+00 3.860785e+00 3.824531e+00
#> [151] 3.787884e+00 3.750855e+00 3.713454e+00 3.675694e+00 3.637585e+00
#> [156] 3.599138e+00 3.560365e+00 3.521278e+00 3.481887e+00 3.442205e+00
#> [161] 3.402243e+00 3.362014e+00 3.321529e+00 3.280800e+00 3.239839e+00
#> [166] 3.198658e+00 3.157270e+00 3.115687e+00 3.073920e+00 3.031984e+00
#> [171] 2.989889e+00 2.947648e+00 2.905275e+00 2.862781e+00 2.820179e+00
#> [176] 2.777481e+00 2.734701e+00 2.691851e+00 2.648944e+00 2.605992e+00
#> [181] 2.563009e+00 2.520006e+00 2.476997e+00 2.433994e+00 2.391010e+00
#> [186] 2.348058e+00 2.305151e+00 2.262301e+00 2.219521e+00 2.176824e+00
#> [191] 2.134222e+00 2.091728e+00 2.049354e+00 2.007114e+00 1.965019e+00
#> [196] 1.923082e+00 1.881316e+00 1.839733e+00 1.798345e+00 1.757164e+00
#> [201] 1.716203e+00 1.675474e+00 1.634988e+00 1.594759e+00 1.554797e+00
#> [206] 1.515115e+00 1.475725e+00 1.436637e+00 1.397865e+00 1.359418e+00
#> [211] 1.321309e+00 1.283548e+00 1.246148e+00 1.209119e+00 1.172472e+00
#> [216] 1.136217e+00 1.100367e+00 1.064931e+00 1.029919e+00 9.953428e-01
#> [221] 9.612119e-01 9.275366e-01 8.943268e-01 8.615923e-01 8.293429e-01
#> [226] 7.975880e-01 7.663372e-01 7.355997e-01 7.053846e-01 6.757008e-01
#> [231] 6.465571e-01 6.179622e-01 5.899246e-01 5.624525e-01 5.355541e-01
#> [236] 5.092374e-01 4.835101e-01 4.583799e-01 4.338543e-01 4.099404e-01
#> [241] 3.866454e-01 3.639762e-01 3.419396e-01 3.205419e-01 2.997897e-01
#> [246] 2.796889e-01 2.602457e-01 2.414657e-01 2.233544e-01 2.059174e-01
#> [251] 1.891597e-01 1.730863e-01 1.577020e-01 1.430114e-01 1.290187e-01
#> [256] 1.157281e-01 1.031436e-01 9.126888e-02 8.010749e-02 6.966272e-02
#> [261] 5.993767e-02 5.093522e-02 4.265803e-02 3.510857e-02 2.828906e-02
#> [266] 2.220152e-02 1.684777e-02 1.222939e-02 8.347738e-03 5.203976e-03
#> [271] 2.799032e-03 1.133618e-03 2.082285e-04 2.313678e-05 5.783981e-04
#> [276] 1.873848e-03 3.909102e-03 6.683558e-03 1.019639e-02 1.444657e-02
#> [281] 1.943282e-02 2.515367e-02 3.160743e-02 3.879219e-02 4.670581e-02
#> [286] 5.534595e-02 6.471005e-02 7.479533e-02 8.559881e-02 9.711728e-02
#> [291] 1.093473e-01 1.222854e-01 1.359275e-01 1.502697e-01 1.653078e-01
#> [296] 1.810372e-01 1.974533e-01 2.145513e-01 2.323261e-01 2.507724e-01
#> [301] 2.698847e-01 2.896575e-01 3.100847e-01 3.311605e-01 3.528784e-01
#> [306] 3.752322e-01 3.982151e-01 4.218204e-01 4.460411e-01 4.708699e-01
#> [311] 4.962996e-01 5.223225e-01 5.489311e-01 5.761173e-01 6.038732e-01
#> [316] 6.321906e-01 6.610609e-01 6.904757e-01 7.204263e-01 7.509038e-01
#> [321] 7.818991e-01 8.134030e-01 8.454063e-01 8.778995e-01 9.108729e-01
#> [326] 9.443167e-01 9.782210e-01 1.012576e+00 1.047371e+00 1.082596e+00
#> [331] 1.118241e+00 1.154295e+00 1.190747e+00 1.227586e+00 1.264802e+00
#> [336] 1.302384e+00 1.340320e+00 1.378600e+00 1.417211e+00 1.456143e+00
#> [341] 1.495383e+00 1.534921e+00 1.574744e+00 1.614841e+00 1.655200e+00
#> [346] 1.695809e+00 1.736655e+00 1.777728e+00 1.819013e+00 1.860501e+00
#> [351] 1.902177e+00 1.944030e+00 1.986047e+00 2.028217e+00 2.070525e+00
#> [356] 2.112960e+00 2.155510e+00 2.198161e+00 2.240902e+00 2.283718e+00
#> [361] 2.326598e+00 2.369529e+00 2.412499e+00 2.455494e+00 2.498501e+00
#> 
#> [[2]]
#>   [1] 4.999657e+00 5.084237e+00 5.168766e+00 5.253220e+00 5.337574e+00
#>   [6] 5.421803e+00 5.505882e+00 5.589785e+00 5.673489e+00 5.756968e+00
#>  [11] 5.840197e+00 5.923152e+00 6.005809e+00 6.088142e+00 6.170128e+00
#>  [16] 6.251742e+00 6.332959e+00 6.413757e+00 6.494110e+00 6.573996e+00
#>  [21] 6.653390e+00 6.732268e+00 6.810609e+00 6.888387e+00 6.965581e+00
#>  [26] 7.042168e+00 7.118124e+00 7.193427e+00 7.268055e+00 7.341986e+00
#>  [31] 7.415197e+00 7.487668e+00 7.559377e+00 7.630302e+00 7.700422e+00
#>  [36] 7.769717e+00 7.838166e+00 7.905749e+00 7.972445e+00 8.038236e+00
#>  [41] 8.103101e+00 8.167022e+00 8.229979e+00 8.291953e+00 8.352927e+00
#>  [46] 8.412883e+00 8.471801e+00 8.529666e+00 8.586460e+00 8.642166e+00
#>  [51] 8.696767e+00 8.750248e+00 8.802593e+00 8.853785e+00 8.903810e+00
#>  [56] 8.952653e+00 9.000300e+00 9.046737e+00 9.091949e+00 9.135923e+00
#>  [61] 9.178647e+00 9.220107e+00 9.260291e+00 9.299188e+00 9.336786e+00
#>  [66] 9.373074e+00 9.408040e+00 9.441675e+00 9.473969e+00 9.504912e+00
#>  [71] 9.534495e+00 9.562709e+00 9.589546e+00 9.614998e+00 9.639056e+00
#>  [76] 9.661716e+00 9.682968e+00 9.702808e+00 9.721229e+00 9.738226e+00
#>  [81] 9.753794e+00 9.767927e+00 9.780623e+00 9.791877e+00 9.801686e+00
#>  [86] 9.810047e+00 9.816958e+00 9.822416e+00 9.826419e+00 9.828968e+00
#>  [91] 9.830060e+00 9.829696e+00 9.827876e+00 9.824599e+00 9.819868e+00
#>  [96] 9.813684e+00 9.806048e+00 9.796963e+00 9.786431e+00 9.774455e+00
#> [101] 9.761040e+00 9.746189e+00 9.729906e+00 9.712196e+00 9.693065e+00
#> [106] 9.672518e+00 9.650561e+00 9.627202e+00 9.602445e+00 9.576300e+00
#> [111] 9.548774e+00 9.519874e+00 9.489610e+00 9.457991e+00 9.425025e+00
#> [116] 9.390723e+00 9.355094e+00 9.318150e+00 9.279901e+00 9.240359e+00
#> [121] 9.199535e+00 9.157442e+00 9.114091e+00 9.069497e+00 9.023671e+00
#> [126] 8.976627e+00 8.928380e+00 8.878944e+00 8.828334e+00 8.776563e+00
#> [131] 8.723649e+00 8.669606e+00 8.614450e+00 8.558198e+00 8.500867e+00
#> [136] 8.442473e+00 8.383034e+00 8.322567e+00 8.261090e+00 8.198622e+00
#> [141] 8.135181e+00 8.070786e+00 8.005455e+00 7.939209e+00 7.872067e+00
#> [146] 7.804048e+00 7.735174e+00 7.665464e+00 7.594938e+00 7.523619e+00
#> [151] 7.451527e+00 7.378683e+00 7.305109e+00 7.230827e+00 7.155858e+00
#> [156] 7.080226e+00 7.003952e+00 6.927059e+00 6.849570e+00 6.771507e+00
#> [161] 6.692895e+00 6.613756e+00 6.534113e+00 6.453990e+00 6.373412e+00
#> [166] 6.292401e+00 6.210983e+00 6.129180e+00 6.047017e+00 5.964519e+00
#> [171] 5.881710e+00 5.798615e+00 5.715258e+00 5.631664e+00 5.547857e+00
#> [176] 5.463863e+00 5.379706e+00 5.295411e+00 5.211004e+00 5.126509e+00
#> [181] 5.041952e+00 4.957357e+00 4.872749e+00 4.788154e+00 4.703596e+00
#> [186] 4.619101e+00 4.534694e+00 4.450400e+00 4.366243e+00 4.282249e+00
#> [191] 4.198442e+00 4.114848e+00 4.031491e+00 3.948395e+00 3.865586e+00
#> [196] 3.783088e+00 3.700926e+00 3.619123e+00 3.537704e+00 3.456694e+00
#> [201] 3.376115e+00 3.295993e+00 3.216350e+00 3.137211e+00 3.058598e+00
#> [206] 2.980536e+00 2.903047e+00 2.826154e+00 2.749880e+00 2.674247e+00
#> [211] 2.599279e+00 2.524997e+00 2.451423e+00 2.378579e+00 2.306487e+00
#> [216] 2.235167e+00 2.164642e+00 2.094932e+00 2.026057e+00 1.958039e+00
#> [221] 1.890897e+00 1.824650e+00 1.759320e+00 1.694925e+00 1.631484e+00
#> [226] 1.569016e+00 1.507539e+00 1.447072e+00 1.387633e+00 1.329239e+00
#> [231] 1.271907e+00 1.215656e+00 1.160500e+00 1.106457e+00 1.053542e+00
#> [236] 1.001772e+00 9.511613e-01 9.017252e-01 8.534783e-01 8.064349e-01
#> [241] 7.606091e-01 7.160142e-01 6.726637e-01 6.305702e-01 5.897464e-01
#> [246] 5.502042e-01 5.119554e-01 4.750113e-01 4.393830e-01 4.050808e-01
#> [251] 3.721151e-01 3.404955e-01 3.102315e-01 2.813320e-01 2.538056e-01
#> [256] 2.276603e-01 2.029041e-01 1.795442e-01 1.575875e-01 1.370405e-01
#> [261] 1.179094e-01 1.001998e-01 8.391690e-02 6.906559e-02 5.565024e-02
#> [266] 4.367484e-02 3.314294e-02 2.405765e-02 1.642167e-02 1.023726e-02
#> [271] 5.506257e-03 2.230054e-03 4.096271e-04 4.551469e-05 1.137825e-03
#> [276] 3.686234e-03 7.689988e-03 1.314790e-02 2.005835e-02 2.841929e-02
#> [281] 3.822825e-02 4.948232e-02 6.217816e-02 7.631201e-02 9.187969e-02
#> [286] 1.088766e-01 1.272976e-01 1.471374e-01 1.683900e-01 1.910492e-01
#> [291] 2.151081e-01 2.405598e-01 2.673966e-01 2.956107e-01 3.251935e-01
#> [296] 3.561365e-01 3.884303e-01 4.220655e-01 4.570320e-01 4.933196e-01
#> [301] 5.309174e-01 5.698143e-01 6.099988e-01 6.514590e-01 6.941826e-01
#> [306] 7.381569e-01 7.833690e-01 8.298053e-01 8.774522e-01 9.262955e-01
#> [311] 9.763207e-01 1.027513e+00 1.079857e+00 1.133338e+00 1.187940e+00
#> [316] 1.243646e+00 1.300439e+00 1.358304e+00 1.417223e+00 1.477178e+00
#> [321] 1.538152e+00 1.600127e+00 1.663084e+00 1.727004e+00 1.791869e+00
#> [326] 1.857660e+00 1.924357e+00 1.991940e+00 2.060389e+00 2.129684e+00
#> [331] 2.199804e+00 2.270729e+00 2.342437e+00 2.414908e+00 2.488120e+00
#> [336] 2.562051e+00 2.636679e+00 2.711982e+00 2.787938e+00 2.864524e+00
#> [341] 2.941718e+00 3.019497e+00 3.097837e+00 3.176716e+00 3.256110e+00
#> [346] 3.335995e+00 3.416349e+00 3.497146e+00 3.578364e+00 3.659978e+00
#> [351] 3.741964e+00 3.824297e+00 3.906953e+00 3.989909e+00 4.073138e+00
#> [356] 4.156617e+00 4.240320e+00 4.324224e+00 4.408303e+00 4.492531e+00
#> [361] 4.576885e+00 4.661339e+00 4.745869e+00 4.830448e+00 4.915053e+00
#> 
#> [[3]]
#>   [1] 1.341877e-01 1.364577e-01 1.387264e-01 1.409931e-01 1.432572e-01
#>   [6] 1.455178e-01 1.477744e-01 1.500263e-01 1.522729e-01 1.545134e-01
#>  [11] 1.567472e-01 1.589737e-01 1.611921e-01 1.634019e-01 1.656024e-01
#>  [16] 1.677928e-01 1.699727e-01 1.721412e-01 1.742978e-01 1.764419e-01
#>  [21] 1.785728e-01 1.806899e-01 1.827925e-01 1.848800e-01 1.869518e-01
#>  [26] 1.890074e-01 1.910460e-01 1.930671e-01 1.950700e-01 1.970543e-01
#>  [31] 1.990192e-01 2.009643e-01 2.028889e-01 2.047925e-01 2.066745e-01
#>  [36] 2.085343e-01 2.103715e-01 2.121853e-01 2.139754e-01 2.157412e-01
#>  [41] 2.174822e-01 2.191977e-01 2.208875e-01 2.225508e-01 2.241873e-01
#>  [46] 2.257965e-01 2.273778e-01 2.289309e-01 2.304552e-01 2.319503e-01
#>  [51] 2.334158e-01 2.348512e-01 2.362561e-01 2.376300e-01 2.389727e-01
#>  [56] 2.402836e-01 2.415624e-01 2.428087e-01 2.440222e-01 2.452024e-01
#>  [61] 2.463491e-01 2.474619e-01 2.485404e-01 2.495844e-01 2.505935e-01
#>  [66] 2.515674e-01 2.525059e-01 2.534086e-01 2.542754e-01 2.551059e-01
#>  [71] 2.558999e-01 2.566571e-01 2.573774e-01 2.580605e-01 2.587062e-01
#>  [76] 2.593144e-01 2.598848e-01 2.604173e-01 2.609117e-01 2.613679e-01
#>  [81] 2.617857e-01 2.621651e-01 2.625058e-01 2.628079e-01 2.630711e-01
#>  [86] 2.632955e-01 2.634810e-01 2.636275e-01 2.637349e-01 2.638033e-01
#>  [91] 2.638327e-01 2.638229e-01 2.637740e-01 2.636861e-01 2.635591e-01
#>  [96] 2.633931e-01 2.631882e-01 2.629443e-01 2.626617e-01 2.623403e-01
#> [101] 2.619802e-01 2.615816e-01 2.611446e-01 2.606693e-01 2.601558e-01
#> [106] 2.596043e-01 2.590150e-01 2.583881e-01 2.577236e-01 2.570219e-01
#> [111] 2.562831e-01 2.555075e-01 2.546952e-01 2.538465e-01 2.529618e-01
#> [116] 2.520411e-01 2.510849e-01 2.500933e-01 2.490667e-01 2.480055e-01
#> [121] 2.469098e-01 2.457800e-01 2.446165e-01 2.434196e-01 2.421897e-01
#> [126] 2.409271e-01 2.396321e-01 2.383053e-01 2.369469e-01 2.355575e-01
#> [131] 2.341373e-01 2.326868e-01 2.312064e-01 2.296967e-01 2.281579e-01
#> [136] 2.265907e-01 2.249954e-01 2.233725e-01 2.217225e-01 2.200459e-01
#> [141] 2.183432e-01 2.166148e-01 2.148614e-01 2.130834e-01 2.112813e-01
#> [146] 2.094558e-01 2.076072e-01 2.057362e-01 2.038434e-01 2.019292e-01
#> [151] 1.999943e-01 1.980392e-01 1.960645e-01 1.940709e-01 1.920588e-01
#> [156] 1.900288e-01 1.879817e-01 1.859179e-01 1.838382e-01 1.817430e-01
#> [161] 1.796331e-01 1.775091e-01 1.753715e-01 1.732211e-01 1.710584e-01
#> [166] 1.688841e-01 1.666989e-01 1.645033e-01 1.622982e-01 1.600840e-01
#> [171] 1.578614e-01 1.556312e-01 1.533939e-01 1.511503e-01 1.489010e-01
#> [176] 1.466467e-01 1.443879e-01 1.421255e-01 1.398601e-01 1.375923e-01
#> [181] 1.353228e-01 1.330523e-01 1.307815e-01 1.285110e-01 1.262416e-01
#> [186] 1.239738e-01 1.217083e-01 1.194459e-01 1.171872e-01 1.149329e-01
#> [191] 1.126836e-01 1.104399e-01 1.082027e-01 1.059725e-01 1.037499e-01
#> [196] 1.015357e-01 9.933053e-02 9.713499e-02 9.494976e-02 9.277549e-02
#> [201] 9.061282e-02 8.846238e-02 8.632482e-02 8.420077e-02 8.209086e-02
#> [206] 7.999571e-02 7.791595e-02 7.585219e-02 7.380505e-02 7.177512e-02
#> [211] 6.976302e-02 6.776933e-02 6.579465e-02 6.383957e-02 6.190465e-02
#> [216] 5.999049e-02 5.809764e-02 5.622666e-02 5.437811e-02 5.255253e-02
#> [221] 5.075048e-02 4.897247e-02 4.721905e-02 4.549072e-02 4.378800e-02
#> [226] 4.211139e-02 4.046140e-02 3.883851e-02 3.724320e-02 3.567594e-02
#> [231] 3.413720e-02 3.262743e-02 3.114709e-02 2.969661e-02 2.827641e-02
#> [236] 2.688693e-02 2.552857e-02 2.420174e-02 2.290682e-02 2.164421e-02
#> [241] 2.041427e-02 1.921737e-02 1.805387e-02 1.692411e-02 1.582842e-02
#> [246] 1.476714e-02 1.374056e-02 1.274901e-02 1.179276e-02 1.087212e-02
#> [251] 9.987336e-03 9.138687e-03 8.326419e-03 7.550775e-03 6.811982e-03
#> [256] 6.110261e-03 5.445819e-03 4.818853e-03 4.229549e-03 3.678082e-03
#> [261] 3.164615e-03 2.689299e-03 2.252277e-03 1.853677e-03 1.493618e-03
#> [266] 1.172205e-03 8.895358e-04 6.456924e-04 4.407474e-04 2.747618e-04
#> [271] 1.477845e-04 5.985326e-05 1.099413e-05 1.221586e-06 3.053851e-05
#> [276] 9.893622e-05 2.063944e-04 3.528814e-04 5.383535e-04 7.627560e-04
#> [281] 1.026022e-03 1.328074e-03 1.668823e-03 2.048167e-03 2.465993e-03
#> [286] 2.922179e-03 3.416589e-03 3.949076e-03 4.519482e-03 5.127640e-03
#> [291] 5.773368e-03 6.456474e-03 7.176758e-03 7.934005e-03 8.727990e-03
#> [296] 9.558479e-03 1.042523e-02 1.132797e-02 1.226645e-02 1.324039e-02
#> [301] 1.424949e-02 1.529346e-02 1.637199e-02 1.748475e-02 1.863143e-02
#> [306] 1.981167e-02 2.102513e-02 2.227145e-02 2.355027e-02 2.486119e-02
#> [311] 2.620384e-02 2.757781e-02 2.898270e-02 3.041809e-02 3.188356e-02
#> [316] 3.337867e-02 3.490298e-02 3.645603e-02 3.803738e-02 3.964654e-02
#> [321] 4.128304e-02 4.294640e-02 4.463612e-02 4.635171e-02 4.809265e-02
#> [326] 4.985844e-02 5.164853e-02 5.346241e-02 5.529954e-02 5.715938e-02
#> [331] 5.904136e-02 6.094494e-02 6.286955e-02 6.481462e-02 6.677958e-02
#> [336] 6.876384e-02 7.076681e-02 7.278790e-02 7.482651e-02 7.688203e-02
#> [341] 7.895387e-02 8.104140e-02 8.314401e-02 8.526107e-02 8.739195e-02
#> [346] 8.953603e-02 9.169266e-02 9.386122e-02 9.604105e-02 9.823151e-02
#> [351] 1.004320e-01 1.026417e-01 1.048602e-01 1.070866e-01 1.093205e-01
#> [356] 1.115610e-01 1.138075e-01 1.160595e-01 1.183161e-01 1.205767e-01
#> [361] 1.228407e-01 1.251074e-01 1.273761e-01 1.296462e-01 1.319169e-01
#>