MICRO: Generate Seasonal Emergence (Lambda) for Emerge model of Aquatic Ecology

Generate lambda for all sites.

makeLambda_Micro(aquaPars)

Arguments

aquaPars
a list of the following structure

N: number of aquatic habitats (required)

lambda: number of emerging adult females per human per day averaged over one year for the entire Landscape (required)

lambdaWeight: vector of weights applied to each site (if not specified or set to NULL initialize to Gamma(1,1) distribution)

offset: vector of seasonal offsets in peak emergence applied to each site (if not specified or set to NULL initialize to 0 for all sites)

Value

list lambda where each element is the daily emergence for that FeedingSite

Examples

makeLambda_Micro(aquaPars= list(lambda = c(2,3,4)))

#> Error in rgamma(n = N, shape = 1, rate = 1): object 'N' not found